fix: MET-182 fix ui bug darkmode
