+++ src/components/StakekeySummary/index.tsx
@@ -192,7 +192,7 @@ const Status = styled("span")<{ status: string }>`
       case "GENERATED":
         return theme.palette.success[100];
       case "IN_PROGRESS":
-        return theme.palette.warning[100];
+        return theme.isDark ? theme.palette.warning[800] : theme.palette.warning[100];
       default:
         return theme.palette.success[100];
     }
@@ -205,7 +205,7 @@ const Status = styled("span")<{ status: string }>`
       case "GENERATED":
         return theme.palette.success[800];
       case "IN_PROGRESS":
-        return theme.palette.warning[800];
+        return theme.isDark ? theme.palette.warning[100] : theme.palette.warning[800];
       default:
         return theme.palette.success[800];
     }
