+++ src/stores/index.tsx
@@ -1,14 +1,14 @@
 import { Store } from "@reduxjs/toolkit";
-import { composeWithDevTools } from "redux-devtools-extension";
 import { applyMiddleware, combineReducers, createStore } from "redux";
+import { composeWithDevTools } from "redux-devtools-extension";
 import { persistReducer, persistStore } from "redux-persist";
 import storage from "redux-persist/lib/storage";
 
-import { RootState } from "./types";
-import userReducer, { setStoreUser } from "./user";
-import userReducer2, { setStoreUser2 } from "./user2";
 import systemReducer, { setStoreSystem } from "./system";
 import toastReducer, { setStoreToast } from "./toast";
+import { RootState } from "./types";
+import userReducer, { setStoreUser } from "./user";
+import themeReducer, { setStoreTheme } from "./theme";
 
 let customStore: Store | undefined;
 
@@ -28,6 +28,12 @@ const userPersistConfig = {
   blacklist: ["onDetailView", "openModal", "modalSignMessage", "modalRegister"]
 };
 
+const themePersistConfig = {
+  key: "theme",
+  storage: storage,
+  blacklist: ["isDark"]
+};
+
 export const getStore = (): Store<RootState> => {
   if (!customStore) {
     throw new Error("Please implement setStore before using this function");
@@ -37,9 +43,9 @@ export const getStore = (): Store<RootState> => {
 
 const appReducer = combineReducers({
   user: persistReducer(userPersistConfig, userReducer),
-  user2: userReducer2,
   system: systemReducer,
-  toast: toastReducer
+  toast: toastReducer,
+  theme: persistReducer(themePersistConfig, themeReducer)
 });
 
 const rootReducer = (state: any, action: any) => appReducer(state, action);
@@ -58,6 +64,6 @@ setStore(store);
 setStoreUser(store);
 setStoreSystem(store);
 setStoreToast(store);
-setStoreUser2(store);
+setStoreTheme(store);
 
 export default store;
