+++ src/types/theme.d.ts
@@ -1,0 +1,5 @@
+export declare type ThemeType = "dark" | "light";
+
+export declare interface ThemeStoreType {
+  theme: ThemeType;
+}
